Back off to a time when 2.11 was still involved
The current version of upickle doesn't include 2.11 (which we still need for a while), and naively trying to add it back fails because some of the dependencies no longer are published for 2.11. So we're backing off to just before 2.11 support was dropped.

For some reason, the Benchmarks aren't compiling. So for now, let's just comment them out -- we don't care about them all that much.

This now fully compiles and passes tests. Note that you must install Mill in order to compile, and Node in order to test.