Progress towards getting things working with 2.13 as well
This code is *horrible*, and desperately needs a good refactoring, but for now I'm focused on making the bloody thing work, first. It's hard to easily refactor, because of how the build.sc file is set up, and the fact that 2.11 and 2.13 have just-plain-contradictory dependencies.

At this point, it's all compiling, but tests are failing.