Tests *mostly* finally appear to pass
Had to ruthlessly pare back bits that we don't care about, and work around some of the weirdnesses of this stack, but we're finally at the point where:
```
mill __.test
```
isn't showing any errors for me.

However, the 2.11 tests are currently commented-out, so this is only partial victory.

Note that this adapts in some of the more-recent changes found in the mainline, after the fork point.